chore: fix breakpoints and naming conventions
