+++ src/models/Events/events.ts
@@ -39,3 +39,12 @@ export interface AllEventResponse {
   previous: number | null;
   results: Event[];
 }
+
+export type TAllEventsPyaload = {
+  event_type: string;
+  is_featured?: boolean;
+  page: number;
+  status: string;
+  search?: string;
+  tag?: string;
+};
