--- src/redux/events/types.ts
@@ -1,8 +1,0 @@
-export type TAllEventsPyaload = {
-  event_type: string;
-  is_featured?: boolean;
-  page: number;
-  status: string;
-  search?: string;
-  tag?: string;
-};
